requirements.txt
cryosparc-tools
python-dotenv
jupyterlab
pytest